+++ rasa/cli/train.py
@@ -14,7 +14,10 @@ def add_subparser(
     import rasa.cli.arguments.train as core_cli
 
     train_parser = subparsers.add_parser(
-        "train", help="Train the Rasa bot", parents=parents
+        "train",
+        help="Train the Rasa bot",
+        parents=parents,
+        formatter_class=argparse.ArgumentDefaultsHelpFormatter,
     )
 
     arguments.train.set_train_arguments(train_parser)
@@ -57,12 +60,14 @@ def train(args: argparse.Namespace) -> Optional[Text]:
     ]
 
     return rasa.train(
-        domain,
-        config,
-        training_files,
-        args.out,
-        args.force,
-        extract_additional_arguments(args),
+        domain=domain,
+        config=config,
+        training_files=training_files,
+        output=args.out,
+        force_training=args.force,
+        fixed_model_name=args.fixed_model_name,
+        uncompress=args.store_uncompressed,
+        kwargs=extract_additional_arguments(args),
     )
 
 
@@ -93,12 +98,14 @@ def train_core(
         config = args.config or DEFAULT_CONFIG_PATH
 
         return train_core(
-            args.domain,
-            config,
-            stories,
-            output,
-            train_path,
-            extract_additional_arguments(args),
+            domain=args.domain,
+            config=config,
+            stories=stories,
+            output=output,
+            train_path=train_path,
+            fixed_model_name=args.fixed_model_name,
+            uncompress=args.store_uncompressed,
+            kwargs=extract_additional_arguments(args),
         )
     else:
         from rasa.core.train import do_compare_training
@@ -119,7 +126,14 @@ def train_nlu(
         args.nlu, "nlu", DEFAULT_DATA_PATH, none_is_valid=True
     )
 
-    return train_nlu(config, nlu_data, output, train_path)
+    return train_nlu(
+        config=config,
+        nlu_data=nlu_data,
+        output=output,
+        train_path=train_path,
+        fixed_model_name=args.fixed_model_name,
+        uncompress=args.store_uncompressed,
+    )
 
 
 def extract_additional_arguments(args: argparse.Namespace) -> Dict: