+++ CHANGELOG.rst
@@ -6,7 +6,7 @@ Rasa Change Log
 All notable changes to this project will be documented in this file.
 This project adheres to `Semantic Versioning`_ starting with version 1.0.
 
-[Unreleased 1.0.0rc1] - `master`_
+[Unreleased 1.0.0rc4] - `master`_
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
 
 Added
@@ -37,6 +37,8 @@ in ``interactive learning``
 - support for spaCy 2.1
 - a model for an agent can now also be loaded from a remote storage
 - log level can be set via environment variable ``LOG_LEVEL``
+- add ``--fixed-model-name`` to ``rasa train`` to specify name of stored model
+- add ``--store-uncompressed`` to ``rasa train`` to not compress Rasa model
 
 Changed
 -------